Scale and translate edges to align with ideal line
Close but not quite, the edges are almost aligned with the ideal line; however the ideal line still drifts off the track and isn't centered between the two edges. Seems to be a scaling issue not a translation issue, if the track was slightly larger then translated it is likely to align.